ref: moved macros into tests itself
