+++ src/flow_store/connection.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 use redis::aio::MultiplexedConnection;
-use redis::{Client, RedisResult};
+use redis::{Client};
 use tokio::sync::Mutex;
 
 pub type FlowStore = Arc<Mutex<Box<MultiplexedConnection>>>;
@@ -29,7 +29,31 @@ mod tests {
     use testcontainers::GenericImage;
     use testcontainers::core::WaitFor;
     use crate::flow_store::connection::build_connection;
-    use crate::redis_container_test;
+
+    macro_rules! redis_container_test {
+        ($test_name:ident, $consumer:expr) => {
+            
+            #[tokio::test]
+            async fn $test_name() {
+                let port: u16 = 6379;
+                let image_name = "redis";
+                let wait_message = "Ready to accept connections";
+    
+                let container = GenericImage::new(image_name, "latest")
+                    .with_exposed_port(port.tcp())
+                    .with_wait_for(WaitFor::message_on_stdout(wait_message))
+                    .start()
+                    .await
+                    .unwrap();
+    
+                let host = container.get_host().await.unwrap();
+                let host_port = container.get_host_port_ipv4(port).await.unwrap();
+                let url = format!("redis://{host}:{host_port}");
+                
+                $consumer(url).await;
+            }
+        };
+    }
     
     redis_container_test!(test_redis_startup, (|url: String| async move {
         println!("Redis server started correctly on: {}", url);
