+++ src/lib.rs
@@ -3,5 +3,3 @@ pub mod flow_store;
 
 #[cfg(feature = "flow_queue")]
 pub mod flow_queue;
-
-pub mod container;
